requirements.txt
flask
flask_restful
tesseract
pytesseract
pillow
requests
PyGithub
markdown
lxml